requirements.txt
openai
feedparser
requests
python-dotenv
groq